Drop the Parser, ParserConfig and Printer traits which only have a single implementation.
