+++ src/parser/md.rs
@@ -26,10 +26,10 @@
 use serde_derive::Deserialize;
 use std::iter::FromIterator;
 
-use super::{ParseError, Parser, ParserConfig, Printer};
+use super::ParseError;
 
 use crate::document::ast::Node;
-use crate::document::code::CodeBlock;
+use crate::document::code::{CodeBlock, Line, Segment, Source};
 use crate::document::text::TextBlock;
 use crate::document::tranclusion::Transclusion;
 use crate::document::Document;
@@ -153,7 +153,7 @@ impl Default for MdParser {
     }
 }
 
-impl<'a> MdParser {
+impl MdParser {
     const LINK_PATTERN: &'static str = r"\[([^\[\]]*)\]\((.*?)\)";
 
     /// Check the validity of the parser configuration
@@ -211,35 +211,33 @@ impl<'a> MdParser {
     }
 }
 
-impl ParserConfig for MdParser {
-    fn comment_start(&self) -> &str {
+impl MdParser {
+    pub fn comment_start(&self) -> &str {
         &self.comment_start
     }
-    fn interpolation_start(&self) -> &str {
+    pub fn interpolation_start(&self) -> &str {
         &self.interpolation_start
     }
-    fn interpolation_end(&self) -> &str {
+    pub fn interpolation_end(&self) -> &str {
         &self.interpolation_end
     }
-    fn macro_start(&self) -> &str {
+    pub fn macro_start(&self) -> &str {
         &self.macro_start
     }
-    fn macro_end(&self) -> &str {
+    pub fn macro_end(&self) -> &str {
         &self.macro_end
     }
-    fn variable_sep(&self) -> &str {
+    pub fn variable_sep(&self) -> &str {
         &self.variable_sep
     }
-    fn file_prefix(&self) -> &str {
+    pub fn file_prefix(&self) -> &str {
         &self.file_prefix
     }
 }
 
-impl Parser for MdParser {
-    type Error = MdError;
-
+impl MdParser {
     #[allow(clippy::nonminimal_bool)]
-    fn parse(&self, input: &str, path: &Path) -> Result<Document, Self::Error> {
+    pub fn parse(&self, input: &str, path: &Path) -> Result<Document, MdError> {
         #[derive(Default)]
         struct State {
             node: Option<Node>,
@@ -428,12 +426,12 @@ impl Parser for MdParser {
         Ok(Document::from_iter(document))
     }
 
-    fn find_links(
+    pub fn find_links(
         &self,
         input: &mut Document,
         from: &Path,
         remove_marker: bool,
-    ) -> Result<Vec<PathBuf>, Self::Error> {
+    ) -> Result<Vec<PathBuf>, MdError> {
         let regex = &self.link_following_pattern;
         let mut paths = vec![];
         let tree = input.tree_mut();
@@ -489,10 +487,136 @@ impl Parser for MdParser {
         }
         Ok(paths)
     }
+
+    /// Parses a macro name, returning the name and the extracted variables
+    #[allow(clippy::type_complexity)]
+    pub fn parse_name(
+        &self,
+        mut input: &str,
+        is_call: bool,
+    ) -> Result<(String, Vec<String>, Vec<Option<String>>), ParseError> {
+        let orig = input;
+        let mut name = String::new();
+        let mut vars = vec![];
+        let mut optionals = vec![];
+        let start = self.interpolation_start();
+        let end = self.interpolation_end();
+        let sep = self.variable_sep();
+        let sep_len = sep.len();
+        loop {
+            if let Some(start_index) = input.find(start) {
+                if let Some(end_index) = input[start_index + start.len()..].find(end) {
+                    name.push_str(&input[..start_index]);
+                    name.push_str(&start);
+                    name.push_str(&end);
+                    let var =
+                        &input[start_index + start.len()..start_index + start.len() + end_index];
+                    if is_call {
+                        vars.push(var.to_owned());
+                        optionals.push(None);
+                    } else if let Some(sep_index) = var.find(sep) {
+                        vars.push((&var[..sep_index]).to_owned());
+                        optionals.push(Some((&var[sep_index + sep_len..]).to_owned()));
+                    } else {
+                        vars.push(var.to_owned());
+                        optionals.push(None);
+                    }
+                    input = &input[start_index + start.len() + end_index + end.len()..];
+                } else {
+                    return Err(ParseError::UnclosedVariableError(orig.to_owned()));
+                }
+            } else {
+                name.push_str(input);
+                break;
+            }
+        }
+        Ok((name, vars, optionals))
+    }
+
+    /// Parses a line as code, returning the parsed `Line` object
+    pub fn parse_line(&self, line_number: usize, input: &str) -> Result<Line, ParseError> {
+        let orig = input;
+        let indent_len = input
+            .chars()
+            .take_while(|ch| ch.is_whitespace())
+            .collect::<String>()
+            .len();
+        let (indent, rest) = input.split_at(indent_len);
+        let (mut rest, comment) = if let Some(comment_index) = rest.find(self.comment_start()) {
+            let (rest, comment) = rest.split_at(comment_index);
+            (
+                rest,
+                Some((&comment[self.comment_start().len()..]).to_owned()),
+            )
+        } else {
+            (rest, None)
+        };
+
+        if rest.starts_with(self.macro_start()) {
+            if let Some(end_index) = rest.find(self.macro_end()) {
+                let (name, scope, _names) =
+                    self.parse_name(&rest[self.macro_start().len()..end_index].trim(), true)?;
+                return Ok(Line {
+                    line_number,
+                    indent: indent.to_owned(),
+                    source: Source::Macro { name, scope },
+                    comment,
+                });
+            }
+        }
+
+        let mut source = vec![];
+        let start = self.interpolation_start();
+        let end = self.interpolation_end();
+        loop {
+            if let Some(start_index) = rest.find(start) {
+                if let Some(end_index) = rest[start_index + start.len()..].find(end) {
+                    source.push(Segment::Source((&rest[..start_index]).to_owned()));
+                    source.push(Segment::MetaVar(
+                        (&rest[start_index + start.len()..start_index + start.len() + end_index])
+                            .to_owned(),
+                    ));
+                    rest = &rest[start_index + start.len() + end_index + end.len()..];
+                } else {
+                    return Err(ParseError::UnclosedVariableError(orig.to_owned()));
+                }
+            } else {
+                if !rest.is_empty() {
+                    source.push(Segment::Source(rest.to_owned()));
+                }
+                break;
+            }
+        }
+
+        Ok(Line {
+            line_number,
+            indent: indent.to_owned(),
+            source: Source::Source(source),
+            comment,
+        })
+    }
+
+    /// Finds all file-specific entry points
+    pub fn get_entry_points(
+        &self,
+        doc: &Document,
+        language: &Option<&str>,
+    ) -> Vec<(String, String)> {
+        let mut entries = vec![];
+        let pref = self.file_prefix();
+        for (name, _block) in doc.tree().code_blocks(language) {
+            if let Some(name) = name {
+                if let Some(rest) = name.strip_prefix(pref) {
+                    entries.push((name.to_owned(), rest.to_owned()))
+                }
+            }
+        }
+        entries
+    }
 }
 
-impl Printer for MdParser {
-    fn print_code_block(&self, block: &CodeBlock) -> String {
+impl MdParser {
+    pub fn print_code_block(&self, block: &CodeBlock) -> String {
         let fence_sequence = if block.alternative {
             &self.fence_sequence_alt
         } else {
@@ -536,7 +660,7 @@ impl Printer for MdParser {
         output
     }
 
-    fn print_code_block_reverse(
+    pub fn print_code_block_reverse(
         &self,
         block: &CodeBlock,
         alternative: Option<&RevCodeBlock>,
@@ -579,11 +703,11 @@ impl Printer for MdParser {
         output
     }
 
-    fn print_text_block(&self, block: &TextBlock) -> String {
+    pub fn print_text_block(&self, block: &TextBlock) -> String {
         format!("{}\n", block.to_string())
     }
 
-    fn print_transclusion(&self, transclusion: &Transclusion, reverse: bool) -> String {
+    pub fn print_transclusion(&self, transclusion: &Transclusion, reverse: bool) -> String {
         let mut output = String::new();
         if reverse {
             output.push_str(transclusion.original());
@@ -592,6 +716,74 @@ impl Printer for MdParser {
             output.push_str("**WARNING!** Missed/skipped transclusion: ");
             output.push_str(transclusion.file().to_str().unwrap());
             output.push('\n');
+        }
+        output
+    }
+
+    /// Fills a name with its placeholders and defaults
+    pub fn print_name(
+        &self,
+        mut name: String,
+        vars: &[String],
+        defaults: &[Option<String>],
+    ) -> String {
+        let start = self.interpolation_start();
+        let end = self.interpolation_end();
+        let var_placeholder = format!("{}{}", start, end);
+        for (var, default) in vars.iter().zip(defaults) {
+            let mut var_full = var.to_string();
+            if let Some(default) = default {
+                var_full.push_str(self.variable_sep());
+                var_full.push_str(default);
+            }
+            let var_name = format!("{}{}{}", start, var_full, end);
+            name = name.replacen(&var_placeholder, &var_name, 1);
+        }
+        name
+    }
+
+    /// Fills a name with its placeholders
+    pub fn print_macro_call(&self, mut name: String, vars: &[String]) -> String {
+        let start = self.interpolation_start();
+        let end = self.interpolation_end();
+        let var_placeholder = format!("{}{}", start, end);
+        for var in vars {
+            let var_name = format!("{}{}{}", start, var, end);
+            name = name.replacen(&var_placeholder, &var_name, 1);
+        }
+        name
+    }
+
+    /// Prints a line of a code block
+    pub fn print_line(&self, line: &Line, print_comments: bool) -> String {
+        let mut output = line.indent.to_string();
+        match &line.source {
+            Source::Macro { name, scope } => {
+                output.push_str(self.macro_start());
+                if !self.macro_start().ends_with(' ') {
+                    output.push(' ');
+                }
+                output.push_str(&self.print_macro_call(name.clone(), &scope));
+                output.push_str(self.macro_end());
+            }
+            Source::Source(segments) => {
+                for segment in segments {
+                    match segment {
+                        Segment::Source(source) => output.push_str(source),
+                        Segment::MetaVar(name) => {
+                            output.push_str(self.interpolation_start());
+                            output.push_str(&name);
+                            output.push_str(self.interpolation_end());
+                        }
+                    }
+                }
+            }
+        }
+        if print_comments {
+            if let Some(comment) = &line.comment {
+                output.push_str(self.comment_start());
+                output.push_str(&comment);
+            }
         }
         output
     }
